// migrations/snapshots/dd5ccf355dd35382111c7a3ee2e13c5eb1fc118e465150aeace8703001caf152/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Answer: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly submissionId: CodecTypes['pg/text@1']['output'];
      readonly fieldId: CodecTypes['pg/text@1']['output'];
      readonly value: CodecTypes['pg/text@1']['output'];
    };
    readonly Form: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly FormField: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly formId: CodecTypes['pg/text@1']['output'];
      readonly label: CodecTypes['pg/text@1']['output'];
      readonly type: CodecTypes['pg/text@1']['output'];
      readonly required: CodecTypes['pg/bool@1']['output'];
      readonly position: CodecTypes['pg/int4@1']['output'];
      readonly options: CodecTypes['pg/json@1']['output'] | null;
      readonly minLength: CodecTypes['pg/int4@1']['output'] | null;
      readonly maxLength: CodecTypes['pg/int4@1']['output'] | null;
      readonly minValue: CodecTypes['pg/float8@1']['output'] | null;
      readonly maxValue: CodecTypes['pg/float8@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
    readonly Submission: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly formId: CodecTypes['pg/text@1']['output'];
      readonly submittedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Answer: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly submissionId: CodecTypes['pg/text@1']['input'];
      readonly fieldId: CodecTypes['pg/text@1']['input'];
      readonly value: CodecTypes['pg/text@1']['input'];
    };
    readonly Form: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly FormField: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly formId: CodecTypes['pg/text@1']['input'];
      readonly label: CodecTypes['pg/text@1']['input'];
      readonly type: CodecTypes['pg/text@1']['input'];
      readonly required: CodecTypes['pg/bool@1']['input'];
      readonly position: CodecTypes['pg/int4@1']['input'];
      readonly options: CodecTypes['pg/json@1']['input'] | null;
      readonly minLength: CodecTypes['pg/int4@1']['input'] | null;
      readonly maxLength: CodecTypes['pg/int4@1']['input'] | null;
      readonly minValue: CodecTypes['pg/float8@1']['input'] | null;
      readonly maxValue: CodecTypes['pg/float8@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
    readonly Submission: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly formId: CodecTypes['pg/text@1']['input'];
      readonly submittedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly answer: {
      readonly fieldId: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly submissionId: CodecTypes['pg/text@1']['output'];
      readonly value: CodecTypes['pg/text@1']['output'];
    };
    readonly form: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly slug: CodecTypes['pg/text@1']['output'];
      readonly title: CodecTypes['pg/text@1']['output'];
    };
    readonly formField: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly formId: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly label: CodecTypes['pg/text@1']['output'];
      readonly maxLength: CodecTypes['pg/int4@1']['output'] | null;
      readonly maxValue: CodecTypes['pg/float8@1']['output'] | null;
      readonly minLength: CodecTypes['pg/int4@1']['output'] | null;
      readonly minValue: CodecTypes['pg/float8@1']['output'] | null;
      readonly options: CodecTypes['pg/json@1']['output'] | null;
      readonly position: CodecTypes['pg/int4@1']['output'];
      readonly required: CodecTypes['pg/bool@1']['output'];
      readonly type: CodecTypes['pg/text@1']['output'];
    };
    readonly submission: {
      readonly formId: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly submittedAt: CodecTypes['pg/timestamptz-temporal@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly answer: {
      readonly fieldId: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly submissionId: CodecTypes['pg/text@1']['input'];
      readonly value: CodecTypes['pg/text@1']['input'];
    };
    readonly form: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly slug: CodecTypes['pg/text@1']['input'];
      readonly title: CodecTypes['pg/text@1']['input'];
    };
    readonly formField: {
      readonly createdAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly formId: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly label: CodecTypes['pg/text@1']['input'];
      readonly maxLength: CodecTypes['pg/int4@1']['input'] | null;
      readonly maxValue: CodecTypes['pg/float8@1']['input'] | null;
      readonly minLength: CodecTypes['pg/int4@1']['input'] | null;
      readonly minValue: CodecTypes['pg/float8@1']['input'] | null;
      readonly options: CodecTypes['pg/json@1']['input'] | null;
      readonly position: CodecTypes['pg/int4@1']['input'];
      readonly required: CodecTypes['pg/bool@1']['input'];
      readonly type: CodecTypes['pg/text@1']['input'];
    };
    readonly submission: {
      readonly formId: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly submittedAt: CodecTypes['pg/timestamptz-temporal@1']['input'];
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;